Clean up & Restore
