--- .github/workflows/main.yml
@@ -1,48 +1,0 @@
-name: Generate Leaderboard
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: '58 18 * * *' 
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v4.1.1
-
-      - name: Set up JDK 21
-        uses: actions/setup-java@v3
-        with:
-          java-version: '21'
-          distribution: 'temurin' 
-          architecture: x64
-
-      - name: Run CLI
-        run: |
-          java -jar ./out/artifacts/CMRITLeaderboard2025_jar/CMRITLeaderboard2025.jar
-          
-      - name: Archive Generated File
-        run: |
-          mkdir -p ${{ github.workspace }}/LeaderboardsStorage
-          cp -f Leaderboards/* ${{ github.workspace }}/LeaderboardsStorage
-
-      - name: Upload Artifact
-        uses: actions/upload-artifact@v4.0.0
-        with:
-          name: leaderboard-artifact
-          path: ${{ github.workspace }}/LeaderboardsStorage
-
-      - name: Get current date
-        id: date
-        run: echo "::set-output name=date::$(date +'%d-%m-%Y %H:%M:%S')"
-
-      - name: Commit Generated Excel File
-        run: |
-          git config --global user.email "actions@github.com"
-          git config --global user.name "GitHub Actions"
-          git add ${{ github.workspace }}/Leaderboards/*
-          git commit -m "Add generated leaderboard data - ${{ steps.date.outputs.date }}"
-          git pull --rebase
-          git push --force
