--- gecko.sh
@@ -1,20 +1,0 @@
-#!/bin/bash
-# download and install latest geckodriver for linux or mac.
-# required for selenium to drive a firefox browser.
-
-set -e;
-install_dir="/usr/local/bin"
-json=$(curl -s https://api.github.com/repos/mozilla/geckodriver/releases/latest)
-if [[ $(uname) == "Darwin" ]]; then
-    url=$(echo "$json" | jq -r '.assets[].browser_download_url | select(contains("macos"))')
-elif [[ $(uname) == "Linux" ]]; then
-    url=$(echo "$json" | jq -r '.assets[].browser_download_url | select(contains("linux64"))')
-else
-    echo "can't determine OS"
-    exit 1
-fi
-url=$(echo $url | cut -f1 -d" ") # new line
-curl -s -L "$url" | tar -xz
-chmod +x geckodriver
-sudo mv geckodriver "$install_dir"
-echo "installed geckodriver binary in $install_dir"
