+++ src/main/java/org/cmrit/CMRITLeaderboard2025.java
@@ -1,4 +1,4 @@
-package org.cmrit;
+package main.java.org.cmrit;
 
 import com.google.gson.Gson;
 import io.github.bonigarcia.wdm.WebDriverManager;
@@ -929,6 +929,8 @@ public class CMRITLeaderboard2025 {
     private static final int MAX_REQUESTS_PER_SECOND = 2;
     private static final long REQUEST_INTERVAL_MILLIS = 1000 / MAX_REQUESTS_PER_SECOND;
 
+    private static long lastRequestTime = 0;
+
     private static void scrapeLeetcode(ArrayList<User> resultSet) {
         // Scraper logic for Leetcode
         System.out.println("Leetcode scraping in progress...");
@@ -947,6 +949,20 @@ public class CMRITLeaderboard2025 {
         int size = resultSet.size();
 
         for (User user : resultSet) {
+
+            // Rate limiting
+            long currentTime = System.currentTimeMillis();
+            long timeElapsedSinceLastRequest = currentTime - lastRequestTime;
+            if (timeElapsedSinceLastRequest < REQUEST_INTERVAL_MILLIS) {
+                try {
+                    Thread.sleep(REQUEST_INTERVAL_MILLIS - timeElapsedSinceLastRequest);
+                } catch (InterruptedException e) {
+                    Thread.currentThread().interrupt();
+                    throw new RuntimeException("Interrupted while sleeping for rate limiting", e);
+                }
+            }
+            lastRequestTime = System.currentTimeMillis();
+
             String handle = user.getHandle();
             String leetcodeHandle = user.getLeetcodeHandle();
             String encodedLeetcodeHandle = URLEncoder.encode(leetcodeHandle, StandardCharsets.UTF_8);
